fix crash, add a few more shortcuts, and extract key commands extensions in to their own file
